+++ fractal.js
@@ -52,7 +52,7 @@ const theme = require('@frctl/mandelbrot')({
   panels: ['notes', 'component', 'html', 'view', 'context'], // Todo: 'info'
   styles: [
     '/styles/dso.css',
-    '/_highlight.js_styles/styles/github.css'
+    '/_highlight.js_styles/github.css'
   ],
   scripts: [
     'https://cdn.jsdelivr.net/npm/vue',